review controller testing, updating user_id on review to optional, and adding review title column

+++ app/controllers/reviews_controller.rb
@@ -5,13 +5,18 @@ class ReviewsController < ApplicationController
   def create
     @review = Review.new(review_params)
 
-    if @review.save
-      flash[:success] = "Your review has been added successfully"
-      product = @review.product
-      redirect_to product_path(product.id)
-      return
+    if @review.valid?
+      if @current_user && @review.product.user_id == @current_user.id
+        flash[:error] = "You can't review your own product!"
+        redirect_to product_path(@review.product.id)
+        return
+      elsif @review.save
+        flash[:success] = "Your review has been added successfully!"
+        redirect_to product_path(@review.product.id)
+        return
+      end
     else
-      flash.now[:error] = "Something went wrong! Review was not added."
+      flash[:error] = "Something went wrong! Review was not added."
       redirect_to root_path
       return
     end
@@ -21,7 +26,7 @@ class ReviewsController < ApplicationController
     review = Review.find_by(id: params[:id])
     if review
       product = review.product
-      if review.user_id == session[:user_id]
+      if review.user_id && review.user_id == session[:user_id]
         review.destroy
         flash[:success] = "Your review was deleted!"
         redirect_to product_path(product.id)
@@ -41,6 +46,6 @@ class ReviewsController < ApplicationController
   private
 
   def review_params
-    return params.require(:review).permit(:rating, :description, :user_id, :product_id)
+    return params.require(:review).permit(:rating, :title, :description, :user_id, :product_id)
   end
 end
